Moved experiment_exception.py back up a level - too many headaches.
